MINOR: cli/debug: show dev: show capabilities
If haproxy compiled with Linux capabilities support, let's show process
capabilities before applying the configuration and at runtime in 'show dev'
stats command output. This maybe useful for debugging purposes. Especially
in the cases, when process changes its UID and GID to non-priviledged or it
started and run under non-priviledged UID and needed capabilities are set by
administrator on the haproxy binary.