MAJOR: haproxy: move daemon and master-worker forks in init
Let's move master-worker forks in init(). In order not to break '-D' and '-p'
modes, we also need to move in init() daemonization fork and a PID file
creation.

Performing master-worker fork at init() allows simplify the code, as only the
worker process for now parses the configuration. Master process reads its
parameters from the command line and after some initialization stages enters
in its mworker_loop, where it waits for the worker process to exit.

So, thanks to this change, there is no need for master to perform additional
reexecution in order to free some memory, when configuration has been loaded
and worker has started. There is also no need for worker to register atexit
callback, to perform routines from mworker_reexec(), in case if it fails at
some init stage.

This rework makes the code of main() more readable and permits to get rid of
HAPROXY_MWORKER_WAIT_ONLY env variable, as now the master process is always in
the wait mode by default, after it has forked the worker. It does not need
reexec anymore, except when it receives USR2 or HUP sigbnals.

We still set MODE_MWORKER_WAIT in this commit to signal the new state of
master, when it has forked a worker.