MINOR: mworker: move check of -W when -S in init_args
To simplify the code of init(), let's move into init_args() the check of -W
(master-worker mode), when -S is provided via cmd line. mworker_cli_conf list
with master sockets configuration is global and should be not empty, if -S is
given and when we exit argument parsing while loop.