MINOR: mworker: move check of -W when -S in init_args
To simplify the code of init(), let's move into init_args() the check of -W
(master-worker mode), when -S is provided via cmd line.

The MODE_MASTER_CLI flag is used to signal, that a new master CLI config line
was inserted into mworker_cli_conf. So, at the end of command line parsing we
can simply check, that MODE_MWORKER is set, when MODE_MASTER_CLI is already
presented in the arg_mode variable.