REORG: mworker: move master-worker fork in init()
This refactoring allows to simplify 'master-worker' logic. The master process
with this change will fork a worker very early at the initialization stage,
which allows to perform a configuration parsing only for the worker. In reality
only the worker process needs to parse and to apply the whole configuration.

Master process just polls master CLI sockets, watches worker status, catches
its termination state and handles the signals. With this refactoring there is
no longer need for master to perform re-execution after reading the whole
configuration file to free additional memory. And there is no longer need for
worker to register atexit callbacks, which must be executed to free resources,
when worker fails to apply the new configuration.

This refactoring allows to remove redandunt variables and to make the code of
main() more readable.

Following patches will try to gather more 'worker-' and 'master-specific' code
in the dedicated cases of this new fork() switch.