MINOR: haproxy: ram_filelist in deinit, init_args, cfgfiles_expand_directories
This commit prepares to store config files content in memory areas allocated
in RAM. So, let's start to use structures of type ram_filelist in all init and
deinit helpers, where we touch cfg_cfgfiles list instead of previous wordlist
type. ram_filelist is an extended version of wordlist type, which along with
the pointer to the config file name keeps the pointer to a memory area, where
the file's content is stored. So it's sufficient just to change the entry
type in deinit(), cfgfiles_expand_directories() and init_args().