MINOR: cfgparse: add struct cfgfile to represent config in .heap
This and following commits serve to prepare loading configuration files in
.heap, before parsing them, as we may need to parse some parts of configuration
in different moments of the startup sequence. This is a case of the new
master-worker initialization process. Here we need to read at first only the
global and the program sections and only after some steps (forking worker, etc)
the rest of the configuration.

Add a new structure cfgfile to keep configuration files metadata and content,
loaded somewhere in a .heap. Instances of filled cfgfile structures could be
chained in a list, as the order in which they were loaded is important.