MEDIUM: cfgparse-global: call some parsers only in MODE_DISCOVERY
This commit is a part of the series to add a support of discovery mode in the
configuration parser and in initialization sequence.

Some keyword parsers tagged with KWF_DISCOVERY (for example those which parse
runtime modes, poller types, pidfile), should not be called twice, i.e., when
the configuration will be read the second time after the discovery mode.
It's redundant and triggers existed parser's errors in standalone mode. In
master-worker mode, worker process inherits parsed settings from the parent.