MINOR: mworker: adapt the code that sets PROC_O_LEAVING
This commit is a part of series to add a support of 4 primary reload use-cases
for the new master-worker architecture.

When master performs a reexec due to receiving USR2 signal or the "reload"
command via master CLI socket, it should set to an already existed worker the
flag PROC_O_LEAVING. It means that existed worked is marked as the previous one
during the reaload and will be terminated after the reload, when a new worker
will be forked and initialized.

In the previous implementation master process was need to do its reexec
twice (the first time for parsing its configuration and the second time to free
unused ressources). So the logic of setting PROC_O_LEAVING was based on
counting the number of reloads. Now, as being mentioned before, reexec is
performed only once, so this allows to simplify this part.