MINOR: mworker/cli: rename and clean mworker_cli_sockpair_new
This commit is a part of series to add the support of 4 primary reload use-cases
for the new master-worker architecture.

Here let's rename mworker_cli_sockpair_new() to
mworker_cli_global_proxy_new_listener() to outline that this function creates
the GLOBAL proxy, allocates the listener with "master-socket" bind conf and
attaches this listener to this GLOBAL proxy. Listener is bound to ipc_fd[1] of
the sockpair inherited in master and in worker (master CLI sockpair).