MEDIUM: startup: refactor sending oldpids_sig logic
This commit is a part of series to add a support of 4 primary reload use-cases
for the new master-worker architecture.

In master-worker and in standalone mode, when the new process has successfully
parsed its configuration and has bound to sockets, it sends either SIGUSR1 or
SIGTERM to the previous one, to terminate it.

Let's reintroduce the support of this feature in the refactored master-worker
architecture. In order to distinguish the worker process from the standalone
mode let's add the mode MODE_WORKER. It's set for the new worker just after the
fork.

If the process is worker, thus it is supervised by the master and it sends
its "READY" string to master via master CLI. So, it's up to master at this
moment to inspect the process list and to stop the previous worker. Master in
this case also sets HAPROXY_LOAD_SUCCESS env variable and emits a corresponding
log message.

If the process in a standalone mode, we also sure at this point that it is
ready to start receiving data. So, if the nb_oldpids counters was incremented
during command line parsing, current process will send oldpids_sig to the
previous one, as it was before the refactoring.