MINOR: mworker: mworker_reexec: set current_child to existed worker
This commit is a part of series to add a support of 4 primary reload use-cases
for the new master-worker architecture.

Due to the moving master-worker fork into the init() and due to performing the
master reexec only once, the logic of setting PROC_O_LEAVING flag and the
moment, where it was set are changed. Now we set this flag for the existed
worker just before reconstructing the master process command line for calling
then execvp() syscall.

So, we need also to adapt the code, which loops over the processes list to find
the worker, that was forked just before entering to mworker_reexec(). We save
its pointer in the "current_child" variable. This variable is needed to set the
PID of the existed worker as an argumnt to "-sf/-st" options and to pass the
master's end of the CLI sockpair (ipc_fd[0]) as an "-x" option argument. "-x"
option should be added to the master's process command line after the
reexec in order to allow the new forked worker to send "_get_socks"
command to master CLI and to obtain all listening sockets from the previous
worker.